requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

torch==2.4.0
scikit-learn==1.5.1

qdrant-client==1.11.1

langchain==0.2.14
langgraph==0.2.36
langchain-groq==0.1.9

sentence-transformers
supabase
python-dotenv

# policy ingestion helpers
PyPDF2
chonkie
python-docx